test: move test files to the `test` directory
