+++ package.json
@@ -32,7 +32,7 @@
     "format": "prettier --write .",
     "lint": "eslint src",
     "sort": "sort-package-json",
-    "test": "mocha \"src/**/*.test.mjs\""
+    "test": "mocha"
   },
   "dependencies": {
     "chalk": "^4.1.2",
